factored out channels, tags, recordings, timers, events and schedules to separate files, deriving from an Entity base class which provides the "dirty" functionality
