+++ CMakeLists.txt
@@ -25,7 +25,6 @@ set(HTS_SOURCES src/AsyncState.cpp
                 src/HTSPConnection.cpp
                 src/HTSPDemuxer.cpp
                 src/HTSPTypes.h
-                src/HTSPTypes.cpp
                 src/HTSPVFS.cpp
                 src/Settings.h
                 src/Tvheadend.cpp
@@ -35,8 +34,25 @@ set(HTS_SOURCES src/AsyncState.cpp
                 src/AutoRecordings.cpp
                 src/AutoRecordings.h
                 src/xbmc_codec_descriptor.hpp)
+
+set(HTS_SOURCES_TVHEADEND_ENTITY
+                src/tvheadend/entity/AutoRecording.h
+                src/tvheadend/entity/AutoRecording.cpp
+                src/tvheadend/entity/Channel.h
+                src/tvheadend/entity/Entity.h
+                src/tvheadend/entity/Event.h
+                src/tvheadend/entity/Recording.h
+                src/tvheadend/entity/RecordingBase.h
+                src/tvheadend/entity/RecordingBase.cpp
+                src/tvheadend/entity/Schedule.h
+                src/tvheadend/entity/Schedule.cpp
+                src/tvheadend/entity/Tag.h
+                src/tvheadend/entity/Tag.cpp
+                src/tvheadend/entity/TimeRecording.h
+                src/tvheadend/entity/TimeRecording.cpp)
                 
 source_group("Source Files" FILES ${HTS_SOURCES})
+source_group("Source Files\\tvheadend\\entity" FILES ${HTS_SOURCES_TVHEADEND_ENTITY})
 
 # Resource files
 set(HTS_RESOURCES 
@@ -51,6 +67,7 @@ source_group("Resource Files" FILES ${HTS_RESOURCES})
 
 # Combine the file lists
 list(APPEND HTS_SOURCES
+            ${HTS_SOURCES_TVHEADEND_ENTITY}
             ${HTS_RESOURCES})
 
 add_subdirectory(lib/libhts)
